# app/temp_files/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Coffee point",
    "description": "This backend application is built on FastAPI and implements the full logic of cafe management.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/check-lists/end/": {
      "get": {
        "tags": [
          "check_lists"
        ],
        "summary": "Get End Check List",
        "operationId": "get_end_check_list_check_lists_end__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckListOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/items/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Read Items",
        "operationId": "read_items_items__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ItemOut"
                  },
                  "type": "array",
                  "title": "Response Read Items Items  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "items"
        ],
        "summary": "Create Item",
        "operationId": "create_item_items__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/items/{item_id}/": {
      "get": {
        "tags": [
          "items"
        ],
        "summary": "Read Item",
        "operationId": "read_item_items__item_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "items"
        ],
        "summary": "Update Item",
        "operationId": "update_item_items__item_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/": {
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Create Order With Products",
        "operationId": "create_order_with_products_orders__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/orders/shift-orders/{shift_id}/": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Shift Orders",
        "operationId": "get_shift_orders_orders_shift_orders__shift_id___get",
        "parameters": [
          {
            "name": "shift_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Shift Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShiftOrderOut"
                  },
                  "title": "Response Get Shift Orders Orders Shift Orders  Shift Id   Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{order_id}/": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "Get Order",
        "operationId": "get_order_orders__order_id___get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShiftOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "Update Order",
        "operationId": "update_order_orders__order_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/orders/status-update/{order_id}/": {
      "put": {
        "tags": [
          "orders"
        ],
        "summary": "Update Order Status",
        "operationId": "update_order_status_orders_status_update__order_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Get Products",
        "operationId": "get_products_products__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductOut"
                  },
                  "type": "array",
                  "title": "Response Get Products Products  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Create Product",
        "operationId": "create_product_products__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/products/{product_id}/": {
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Update Product",
        "operationId": "update_product_products__product_id___put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Deactivate Product",
        "operationId": "deactivate_product_products__product_id___delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductFullInfoOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "active"
        ],
        "title": "CategoryOut"
      },
...
      "CheckListOut": {
        "properties": {
          "check_list": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Check List"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "check_list",
          "id"
        ],
        "title": "CheckListOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement"
        ],
        "title": "ItemCreate"
      },
      "ItemOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement",
          "id",
          "active"
        ],
        "title": "ItemOut"
      },
      "ItemUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "measurement": {
            "type": "string",
            "title": "Measurement"
          }
        },
        "type": "object",
        "required": [
          "name",
          "measurement"
        ],
        "title": "ItemUpdate"
      },
      "OrderCreate": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "payment_method": {
            "$ref": "#/components/schemas/OrderPaymentMethod"
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "shift_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shift Id"
          },
          "client_id": {
            "title": "Client Id",
            "type": [
              "string",
              "null"
            ]
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/ProductOrderCreate"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "price",
          "date",
          "payment_method",
          "type",
          "status",
          "shift_id",
          "products"
        ],
        "title": "OrderCreate"
      },
      "OrderOut": {
        "properties": {
          "price": {
            "type": "string",
            "title": "Price"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "payment_method": {
            "$ref": "#/components/schemas/OrderPaymentMethod"
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "shift_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shift Id"
          },
          "client_id": {
            "title": "Client Id",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "order_number": {
            "type": "integer",
            "title": "Order Number"
          }
        },
        "type": "object",
        "required": [
          "price",
          "date",
          "payment_method",
          "type",
          "status",
          "shift_id",
          "id",
          "active",
          "order_number"
        ],
        "title": "OrderOut"
      },
      "OrderPaymentMethod": {
        "type": "string",
        "enum": [
          "cash",
          "card"
        ],
        "title": "OrderPaymentMethod"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "completed",
          "cancelled",
          "returned"
        ],
        "title": "OrderStatus"
      },
      "OrderStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OrderStatusUpdate"
      },
      "OrderType": {
        "type": "string",
        "enum": [
          "dine_in",
          "delivery",
          "takeout"
        ],
        "title": "OrderType"
      },
      "OrderUpdate": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "date": {
            "title": "Date",
            "type": [
              "string",
              "null"
            ]
          },
          "payment_method": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderPaymentMethod"
              },
              {
                "type": "null"
              }
            ]
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrderType"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_id": {
            "title": "Client Id",
            "type": [
              "string",
              "null"
            ]
          },
          "products": {
            "title": "Products",
            "type": [
              "array",
              "null"
            ]
          }
        },
        "type": "object",
        "title": "OrderUpdate"
      },
      "ProductCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Price"
          },
          "online_shop": {
            "type": "boolean",
            "title": "Online Shop"
          },
          "image_url": {
            "title": "Image Url",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id",
          "price",
          "online_shop"
        ],
        "title": "ProductCreate"
      },
      "ProductFullInfoOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "online_shop": {
            "type": "boolean",
            "title": "Online Shop"
          },
          "image_url": {
            "title": "Image Url",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category_id",
          "price",
          "online_shop",
          "id",
          "active"
        ],
        "title": "ProductFullInfoOut"
      },
      "ProductOrderCreate": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "count"
        ],
        "title": "ProductOrderCreate"
      },
...
      "ProductOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "online_shop": {
            "type": "boolean",
            "title": "Online Shop"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryOut"
          },
          "image_url": {
            "title": "Image Url",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "price",
          "online_shop",
          "category"
        ],
        "title": "ProductOut"
      },
      "ProductUpdate": {
        "properties": {
          "name": {
            "title": "Name",
            "type": [
              "string",
              "null"
            ]
          },
          "category_id": {
            "title": "Category Id",
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "online_shop": {
            "title": "Online Shop",
            "type": [
              "boolean",
              "null"
            ]
          },
          "image_url": {
            "title": "Image Url",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object",
        "title": "ProductUpdate"
      },
...
      "ShiftOrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "payment_method": {
            "$ref": "#/components/schemas/OrderPaymentMethod"
          },
          "type": {
            "$ref": "#/components/schemas/OrderType"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "client_id": {
            "title": "Client Id",
            "type": [
              "string",
              "null"
            ]
          },
          "order_number": {
            "type": "integer",
            "title": "Order Number"
          },
          "products": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__product__ProductOrderOut"
            },
            "type": "array",
            "title": "Products"
          }
        },
        "type": "object",
        "required": [
          "id",
          "active",
          "price",
          "date",
          "payment_method",
          "type",
          "status",
          "order_number",
          "products"
        ],
        "title": "ShiftOrderOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}